Dialogue System Script
Added a script that allows for a dialogue system when the 'E' key is pressed. Needs to be altered to only popup when player is in a certain trigger, but for now works whenever the key is pressed. Also added a test scene that holds the UI for the dialogue.