DialogueTree script added and implemented
Added a new script called DialogueTree that acts the same as the other interaction scripts, but this one brings up a list of questions to click. Each question has different dialogue and when that's done the question list will pop up again. There is also a close button that will end the dialogue and make the questions go away.