Fixed multiple issues with the Interaction/Dialogue system
Fixed an issue where the dialogue of the last interactable object that the player encountered would pop up when the interact button was pressed no matter where the player was. This was caused by the character controlller not emptying out its currentInteractable variable which caused the function that checks for it to always run since it was always true.

Also fixed an issue in the DialogueTree and MenuInteraction scripts where once the UI popped up, if the player kept pressing the interact key then the dialogue would pop up again.