Dialogue now includes a name
The interaction scripts now take a DialogueWithName asset and use that to pass through to DialogueManager. DialogueManager was also changed so that its function now take a DialogueWithName object. Also added a new image and text box to the Dialogue UI so that the name string shows while in dialogue