Interact script with custom editor added and implemented
Added an Interact script that holds a custom editor so that it can be easily edited in the properties menu. This script is also a combination of all the other interact scripts rolled up into one. Due to the custom editor you can select what interaction type that you want and the appropriate fields will pop up.