Interact updated to lock cursor appropriately
Updated the Interact script so that it unlocks the cursor when the menu interact pops up its UI. Also added a back button so that the UI can be closed and the game doesn't have to be entered. Also altered the item interaction type so that when the item is collected it gets added to a list in the Game State object.
A new script called DontRespawn will access that list and if the object is on there on Start then it will be deleted to simulate already being collected.