Updated interact to include a story trigger
Updated the interact script so that it now includes a bool to see if the interact needs to be in a certain story state so that the interact can work. The state that it check is a enum in GameState that will update as needed so interact points will become available as needed.