Added Hearts UI to go along with player health and Also implemented invic frames
I added the hearts UI the player can right now have a max hearts of 10 and can go down from there. There is now invic frames when the player is hit they can will no longer be able to be hit until a certain amount of time as passed and also the player blinks to indicate they are invincable. Still need to implement changes to other cyber worlds