Updated gun rotation on keyboard to go with keyboard
Now you use w to aim up and use d and w or a and w to aim up. I also started the free lock it only allows you to shoot in the way you are facing may change it may not. Adjusted some camera values and upped the players speed