Added Reticle and updated some levels
The gamemanger has the two images and bools to determine which game it is in to set the curors. Health pickup now uses the value of healthToGive to display the number with the text. Level designed the tut a little better and just added decorations a bit also the levels are in the build