Fixed volume slider issue
This will need to be copied over into the main menu as well. I am unable to edit the prefab there as everything is grayed out for me now (maybe locked for editing).  To fix this issue bring the fill part in slightly on the right and left so that when you move the slider the fill bar doesnt go beyound the fill area.