Working on Settings > controls page
Currently it is not switching between the two UI panels as expected. I have tried having it auto detect the panels via the game manager and also using a separate script (verify controller), but neither case brought up the panels at the correct time. Futher research will need to be done on how to get these panels to work.